screencapture/requirements.txt
mss==9.0.1
pillow==10.3.0
